requirements.txt
APScheduler==3.3.1
asn1crypto==0.22.0
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
cffi==1.9.1
chardet==2.3.0
click==6.7
cryptography==2.0.2
cssselect==0.9.2
enum34==1.1.6
Flask==0.11.1
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.7.1
html5lib==0.999
idna==2.5
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.2
MarkupSafe==0.23
olefile==0.44
Pillow==4.1.1
ply==3.9
pycparser==2.14
pyOpenSSL==16.2.0
PySocks==1.5.7
pytz==2016.10
redis==2.10.6
reportlab==3.3.0
requests==2.13.0
rq==0.8.2
six==1.10.0
tzlocal==1.4
urllib3==1.20
Werkzeug==0.11.10
WTForms==2.0
